move test_compose_out_with_color into tests.test_in_out
